Moved output functions to output.py
 - Sorry Nick.
