+++ docs/source/utils.rst
@@ -5,27 +5,9 @@ Utilities
 
 .. contents:: Table of Contents
 
-Command Output
---------------
-
-.. autofunction:: pavilion.utils.fprint
-.. autofunction:: pavilion.utils.draw_table
-.. autofunction:: pavilion.utils.dbg_print
-
-Colorized Output
-~~~~~~~~~~~~~~~~
-Both ``fprint()`` and ``dbg_print`` above take a 'color' argument, which
-allows you to colorize the output.
-
-.. autodata:: pavilion.utils.COLORS
-    :annotation:
-.. autoclass:: pavilion.utils.ANSIString
-    :members:
-
-JSON Output
-~~~~~~~~~~~
-.. autofunction:: pavilion.utils.json_dump
-.. autofunction:: pavilion.utils.json_dump
+Test and Suite Directories
+--------------------------
+.. autofunction:: pavilion.utils.make_id_path
 
 File Handling
 -------------
