+++ lib/pavilion/plugins/commands/cancel.py
@@ -1,13 +1,15 @@
+"""Cancels tests as prescribed by the user."""
+
 import errno
 import os
-import sys
+
 from pavilion import commands
+from pavilion import output
 from pavilion import schedulers
-from pavilion import utils
 from pavilion import series
+from pavilion.plugins.commands.status import print_from_test_obj
 from pavilion.status_file import STATES
 from pavilion.test_run import TestRun, TestRunError
-from pavilion.plugins.commands.status import print_from_test_obj
 
 
 class CancelCommand(commands.Command):
@@ -43,8 +45,9 @@ class CancelCommand(commands.Command):
         )
 
     def run(self, pav_cfg, args):
+        """Cancel the given tests."""
 
-        user_id = os.geteuid() # gets unique user id
+        user_id = os.geteuid()  # gets unique user id
 
         if not args.tests:
             # user wants to cancel all current tests
@@ -71,28 +74,28 @@ class CancelCommand(commands.Command):
                                                                test_id)
                                      .tests)
                 except series.TestSeriesError as err:
-                    utils.fprint(
+                    output.fprint(
                         "Series {} could not be found.\n{}".format(test_id,
                                                                    err),
                         file=self.errfile,
-                        color=utils.RED
+                        color=output.RED
                     )
                     return errno.EINVAL
                 except ValueError as err:
-                    utils.fprint(
+                    output.fprint(
                         "Series {} is not a valid series.\n{}"
                         .format(test_id, err), file=self.errfile,
-                        color=utils.RED
+                        color=output.RED
                     )
                     return errno.EINVAL
             else:
                 try:
                     test_list.append(int(test_id))
                 except ValueError as err:
-                    utils.fprint(
+                    output.fprint(
                         "Test {} is not a valid test.\n{}".format(test_id,
                                                                   err),
-                        file=self.errfile, color=utils.RED
+                        file=self.errfile, color=output.RED
                     )
                     return errno.EINVAL
 
@@ -115,21 +118,24 @@ class CancelCommand(commands.Command):
                     cancel_status = sched.cancel_job(test)
                     test.status.set(cancel_status.state, cancel_status.note)
                     test.set_run_complete()
-                    utils.fprint("Test {} cancelled."
-                                 .format(test_id), file=self.outfile,
-                                 color=utils.GREEN)
+                    output.fprint(
+                        "Test {} cancelled."
+                        .format(test_id), file=self.outfile,
+                        color=output.GREEN)
 
                 else:
-                    utils.fprint("Test {} could not be cancelled has state: {}."
-                                 .format(test_id, status.state),
-                                 file=self.outfile,
-                                 color=utils.RED)
+                    output.fprint(
+                        "Test {} could not be cancelled has state: {}."
+                        .format(test_id, status.state),
+                        file=self.outfile,
+                        color=output.RED)
 
             except TestRunError as err:
-                utils.fprint("Test {} could not be cancelled, cannot be" \
-                             " found. \n{}".format(test_id, err),
-                             file=self.errfile,
-                             color=utils.RED)
+                output.fprint(
+                    "Test {} could not be cancelled, cannot be found. \n{}"
+                    .format(test_id, err),
+                    file=self.errfile,
+                    color=output.RED)
                 return errno.EINVAL
 
         # Only prints statuses of tests if option is selected
