+++ lib/pavilion/plugins/commands/cat.py
@@ -1,11 +1,16 @@
+"""Prints files from a given test run id."""
+
 import errno
 import os
 import sys
 
+from pavilion import output
 from pavilion import commands
 from pavilion import utils
 
+
 class FileCommand(commands.Command):
+    """Prints the given file for a given test run id."""
 
     def __init__(self):
         super().__init__(
@@ -26,34 +31,39 @@ class FileCommand(commands.Command):
         )
 
     def run(self, pav_cfg, args):
+        """Run this command."""
 
         test_dir = pav_cfg.working_dir/'test_runs'
         job_dir = utils.make_id_path(test_dir, args.job_id)
 
         if os.path.isdir(job_dir.as_posix()) is False:
-            utils.fprint("directory '{}' does not exist."
-                         .format(job_dir.as_posix()),
-                         file=sys.stderr, color=utils.RED)
+            output.fprint("directory '{}' does not exist."
+                          .format(job_dir.as_posix()),
+                          file=sys.stderr, color=output.RED)
             return errno.EEXIST
 
         return print_file(job_dir/args.file)
 
 
 def print_file(file):
+    """Print the file at the given path.
+    :param path file: The path to the file to print.
+    """
+
     try:
-        with open(file, 'r') as file:
+        with file.open() as file:
             while True:
                 block = file.read(4096)
                 if not block:
                     break
-                utils.fprint(block, file=sys.stdout)
+                output.fprint(block, file=sys.stdout)
 
     except IsADirectoryError:
-        utils.fprint("{} is a directory.".format(file), sys.stderr,
-                     color=utils.RED)
+        output.fprint("{} is a directory.".format(file), sys.stderr,
+                      color=output.RED)
         return errno.EINVAL
 
     except FileNotFoundError:
-        utils.fprint("file '{}' does not exist.".format(file), sys.stderr,
-                     color=utils.RED)
+        output.fprint("file '{}' does not exist.".format(file), sys.stderr,
+                      color=output.RED)
         return errno.EEXIST
