+++ lib/pavilion/plugins/commands/clean.py
@@ -1,17 +1,17 @@
 import errno
-import sys
-import argparse
 import os
 import shutil
 from calendar import monthrange
 from datetime import datetime, timedelta
+
+from pavilion import output
 from pavilion import commands
-from pavilion import utils
-from pavilion.test_run import TestRun, TestRunError, TestRunNotFoundError
 from pavilion.status_file import STATES
+from pavilion.test_run import TestRun, TestRunError, TestRunNotFoundError
 
 
 class CleanCommand(commands.Command):
+    """Cleans outdated test and series run directories."""
 
     def __init__(self):
         super().__init__(
@@ -28,10 +28,11 @@ class CleanCommand(commands.Command):
         parser.add_argument(
             '--older-than', nargs='+', action='store',
             help='Set the max age of files to be removed. Can be a date ex:'
-            '"Jan 1 2019" or , or a number of days/weeks ex:"32 weeks"'
+                 '"Jan 1 2019" or , or a number of days/weeks ex:"32 weeks"'
         )
 
     def run(self, pav_cfg, args):
+        """Run this command."""
 
         if args.older_than:
             if 'day' in args.older_than or 'days' in args.older_than:
@@ -46,112 +47,113 @@ class CleanCommand(commands.Command):
                 date = ' '.join(args.older_than)
                 try:
                     cutoff_date = datetime.strptime(date, '%b %d %Y')
-                except (TypeError, ValueError) as err:
-                    utils.fprint("{} is not a valid \
-                                 date.".format(args.older_than),
-                                 file=self.errfile, color=utils.RED)
+                except (TypeError, ValueError):
+                    output.fprint("{} is not a valid date."
+                                  .format(args.older_than),
+                                  file=self.errfile, color=output.RED)
                     return errno.EINVAL
 
         # No cutoff specified, removes everything.
         else:
             cutoff_date = datetime.today()
 
-        tests_dir = pav_cfg.working_dir/'test_runs'
-        series_dir = pav_cfg.working_dir/'series'
-        download_dir = pav_cfg.working_dir/'downloads'
-        build_dir = pav_cfg.working_dir/'builds'
+        tests_dir = pav_cfg.working_dir / 'test_runs'
+        series_dir = pav_cfg.working_dir / 'series'
+        download_dir = pav_cfg.working_dir / 'downloads'
+        build_dir = pav_cfg.working_dir / 'builds'
 
         dependent_builds = []
         incomplete_tests = []
         # Clean Tests
-        utils.fprint("Removing Tests...", file=self.outfile,
-                     color=utils.GREEN)
+        output.fprint("Removing Tests...", file=self.outfile,
+                      color=output.GREEN)
         for test in os.listdir(tests_dir.as_posix()):
             test_time = datetime.fromtimestamp(
-                os.path.getmtime((tests_dir/test).as_posix()))
+                os.path.getmtime((tests_dir / test).as_posix()))
             try:
                 test_obj = TestRun.load(pav_cfg, int(test))
                 status = test_obj.status.current().state
-            except (TestRunError, TestRunNotFoundError) as err:
-                utils.fprint("Removing bad test directory {}".format(test),
-                             file=self.outfile)
+            except (TestRunError, TestRunNotFoundError):
+                output.fprint("Removing bad test directory {}".format(test),
+                              file=self.outfile)
                 shutil.rmtree(tests_dir.as_posix())
                 continue
             if test_time < cutoff_date and status != STATES.RUNNING \
-                                       and status != STATES.SCHEDULED:
-                shutil.rmtree((tests_dir/test).as_posix())
+                    and status != STATES.SCHEDULED:
+                shutil.rmtree((tests_dir / test).as_posix())
                 if args.verbose:
-                    utils.fprint("Removed test {}".format(test),
-                                 file=self.outfile)
+                    output.fprint("Removed test {}".format(test),
+                                  file=self.outfile)
             else:
                 if args.verbose:
-                    utils.fprint("Skipped test {}".format(test),
-                                 file=self.outfile)
+                    output.fprint("Skipped test {}".format(test),
+                                  file=self.outfile)
                 incomplete_tests.append(test)
                 dependent_builds.append(test_obj.build_name)
 
         # Clean Series
         completed_series = True
-        utils.fprint("Removing Series...", file=self.outfile,
-                     color=utils.GREEN)
+        output.fprint("Removing Series...", file=self.outfile,
+                      color=output.GREEN)
         for series in os.listdir(series_dir.as_posix()):
             series_time = datetime.fromtimestamp(
-                os.path.getmtime((series_dir/series).as_posix()))
+                os.path.getmtime((series_dir / series).as_posix()))
             for test in incomplete_tests:
-                if os.path.exists((series_dir/series/test).as_posix()):
+                if os.path.exists((series_dir / series / test).as_posix()):
                     completed_series = False
             if series_time < cutoff_date and completed_series:
-                shutil.rmtree((series_dir/series).as_posix())
+                shutil.rmtree((series_dir / series).as_posix())
                 if args.verbose:
-                    utils.fprint("Removed series {}".format(series),
-                                 file=self.outfile)
+                    output.fprint("Removed series {}".format(series),
+                                  file=self.outfile)
             else:
                 if args.verbose:
-                    utils.fprint("Skipped series {}".format(series),
-                                 file=self.outfile)
+                    output.fprint("Skipped series {}".format(series),
+                                  file=self.outfile)
 
         # Clean Downloads
-        utils.fprint("Removing Downloads...", file=self.outfile,
-                     color=utils.GREEN)
+        output.fprint("Removing Downloads...", file=self.outfile,
+                      color=output.GREEN)
         for download in os.listdir(download_dir.as_posix()):
             download_time = datetime.fromtimestamp(
-                os.path.getmtime((download_dir/download).as_posix()))
+                os.path.getmtime((download_dir / download).as_posix()))
             if download_time < cutoff_date:
                 try:
-                    shutil.rmtree((download_dir/download).as_posix())
-                except NotADirectoryError as err:
-                    utils.fprint("{} is not a directory.".format(download),
-                                 file=self.errfile, color=utils.RED)
-                    os.remove((download_dir/download).as_posix())
+                    shutil.rmtree((download_dir / download).as_posix())
+                except NotADirectoryError:
+                    output.fprint("{} is not a directory.".format(download),
+                                  file=self.errfile, color=output.RED)
+                    os.remove((download_dir / download).as_posix())
                 if args.verbose:
-                    utils.fprint("Removed download {}".format(download),
-                                 file=self.outfile)
+                    output.fprint("Removed download {}".format(download),
+                                  file=self.outfile)
             else:
                 if args.verbose:
-                    utils.fprint("Skipped download {}".format(download),
-                                 file=self.outfile)
+                    output.fprint("Skipped download {}".format(download),
+                                  file=self.outfile)
 
         # Clean Builds
-        utils.fprint("Removing Builds...", file=self.outfile, color=utils.GREEN)
+        output.fprint("Removing Builds...", file=self.outfile,
+                      color=output.GREEN)
         for build in os.listdir(build_dir.as_posix()):
             build_time = datetime.fromtimestamp(
-                os.path.getmtime((build_dir/build).as_posix()))
+                os.path.getmtime((build_dir / build).as_posix()))
             if build_time < cutoff_date and build not in dependent_builds:
-                shutil.rmtree((build_dir/build).as_posix())
+                shutil.rmtree((build_dir / build).as_posix())
                 if args.verbose:
-                    utils.fprint("Removed build {}".format(build),
-                                 file=self.outfile)
+                    output.fprint("Removed build {}".format(build),
+                                  file=self.outfile)
             else:
                 if args.verbose:
-                    utils.fprint("Skipped build {}".format(build),
-                                 file=self.outfile)
+                    output.fprint("Skipped build {}".format(build),
+                                  file=self.outfile)
 
         return 0
 
 
 def get_month_delta(months):
+    """Turn a number of months in the future into a concrete date."""
 
-    cutoff_date = None
     today = datetime.today()
     cur_year = today.year
     cur_day = today.day
